requirements.txt
arrow==1.3.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
DateTime==5.5
emailmessage==0.2.1
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
fonttools==4.58.5
fsspec==2025.5.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
MouseInfo==0.1.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
numpy==2.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pillow==11.3.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-time==0.3.0
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
requests==2.32.4
secure-smtplib==0.1.1
selenium==4.34.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wikipedia==1.4.0
wsproto==1.2.0
zope.interface==7.2
gunicorn
psycopg2-binary
